Add get all users route

+++ src/routes/personal.route.ts
@@ -20,7 +20,7 @@ router.get('/thanked-reviews', requireAuth, getPersonalThankedReviews);
 
 router.patch(
   '/',
-  param('id').toInt(),
+  param('id'),
   body('username').optional().notEmpty(),
   body('gender').optional().toLowerCase().isIn(['male', 'female', 'other']),
   body('dateOfBirth')
